updated classrom endpoints

+++ FoodplannerApi/Controller/ClassroomsController.cs
@@ -19,6 +19,7 @@ public class ClassroomsController : BaseController
         return Ok(classroom);
     }
 
+    [Authorize(Policy = "AdminPolicy")]
     [HttpPost]
     public async Task<ActionResult> Create([FromBody] CreateClassroomDTO createClassroomDTO){
         var result = await _classroomService.InsertClassroomAsync(createClassroomDTO);
@@ -28,6 +29,7 @@ public class ClassroomsController : BaseController
         return BadRequest();
     }
 
+    [Authorize(Policy = "AdminPolicy")]
     [HttpPut("{id}")]
     public async Task<ActionResult> Update([FromBody] CreateClassroomDTO createClassroomDTO, int id){
         var result = await _classroomService.UpdateClassroomAsync(createClassroomDTO, id);
@@ -37,8 +39,13 @@ public class ClassroomsController : BaseController
         return BadRequest();
     }
     
+    [Authorize(Policy = "AdminPolicy")]
     [HttpDelete("{id}")]
     public async Task<ActionResult> Delete(int id){
+        var isChildrenInClassroom = await _classroomService.CheckChildrenInClassroom(id);
+        if (isChildrenInClassroom){
+            return BadRequest(new ErrorResponse{Message = ["Der er børn i klassen"]});
+        }
         var result = await _classroomService.DeleteClassroomAsync(id);
         if (result > 0){
             return Ok(result);
